create a popover common component
