+++ client/src/components/Common/DisabledPopOver/DisabledPopOver.style.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+export const ContentWrapper = styled.div`
+  font-family: Roboto;
+  font-style: normal;
+  font-weight: bold;
+  cursor: not-allowed;
+  color: #0ac7e7;
+
+  a {
+    cursor: not-allowed;
+  }
+`;
